Add list of bids

+++ src/main/kotlin/org/kor/portal/service/robofinist/model/program/Program.kt
@@ -3,6 +3,6 @@ package org.kor.portal.service.robofinist.model.program
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class Program(
-    @param:JsonProperty("id") val id: Int,
+    @param:JsonProperty("id") val id: Long,
     @param:JsonProperty("name") val name: String,
 )
